mlua_sys::lua51

Module lua

Contains definitions from lua.h.

Structs

lua_Debug

lua_State

A raw Lua state associated with a thread.

Constants

LUA_ENVIRONINDEX

LUA_ERRERR

LUA_ERRMEM

LUA_ERRRUN

LUA_ERRSYNTAX

LUA_GCCOLLECT

LUA_GCCOUNT

LUA_GCCOUNTB

LUA_GCRESTART

LUA_GCSETPAUSE

LUA_GCSETSTEPMUL

LUA_GCSTEP

LUA_GCSTOP

LUA_GLOBALSINDEX

LUA_HOOKCALL

LUA_HOOKCOUNT

LUA_HOOKLINE

LUA_HOOKRET

LUA_HOOKTAILCALL

LUA_MASKCALL

LUA_MASKCOUNT

LUA_MASKLINE

LUA_MASKRET

LUA_MINSTACK

Minimum Lua stack available to a C function

LUA_MULTRET

LUA_OK

LUA_REGISTRYINDEX

LUA_SIGNATURE

LUA_TBOOLEAN

LUA_TFUNCTION

LUA_TLIGHTUSERDATA

LUA_TNIL

LUA_TNONE

LUA_TNUMBER

LUA_TSTRING

LUA_TTABLE

LUA_TTHREAD

LUA_TUSERDATA

LUA_YIELD

Functions

lua_atpanic⚠

lua_call⚠

lua_checkstack⚠

lua_close⚠

lua_concat⚠

lua_cpcall⚠

lua_createtable⚠

lua_dump_⚠

lua_equal⚠

lua_error⚠

lua_gc⚠

lua_getallocf⚠

lua_getfenv⚠

lua_getfield_⚠

lua_getglobal_⚠

lua_gethook⚠

lua_gethookcount⚠

lua_gethookmask⚠

lua_getinfo⚠

lua_getlocal⚠

lua_getmetatable⚠

lua_getstack⚠

lua_gettable_⚠

lua_gettop⚠

lua_getupvalue⚠

lua_insert⚠

lua_isboolean⚠

lua_iscfunction⚠

lua_isfunction⚠

lua_islightuserdata⚠

lua_isnil⚠

lua_isnone⚠

lua_isnoneornil⚠

lua_isnumber⚠

lua_isstring⚠

lua_istable⚠

lua_isthread⚠

lua_isuserdata⚠

lua_lessthan⚠

lua_load⚠

lua_newstate⚠

lua_newtable⚠

lua_newthread⚠

lua_newuserdata⚠

lua_next⚠

lua_objlen⚠

lua_pcall⚠

lua_pop⚠

lua_pushboolean⚠

lua_pushcclosure⚠

lua_pushcfunction⚠

lua_pushfstring⚠

lua_pushinteger⚠

lua_pushlightuserdata⚠

lua_pushliteral⚠

lua_pushlstring_⚠

lua_pushnil⚠

lua_pushnumber⚠

lua_pushstring_⚠

lua_pushthread⚠

lua_pushvalue⚠

lua_rawequal⚠

lua_rawget_⚠

lua_rawgeti_⚠

lua_rawset⚠

lua_rawseti_⚠

lua_register⚠

lua_remove⚠

lua_replace⚠

lua_resume_⚠

lua_setallocf⚠

lua_setfenv⚠

lua_setfield⚠

lua_setglobal⚠

lua_sethook⚠

lua_setlocal⚠

lua_setmetatable⚠

lua_settable⚠

lua_settop⚠

lua_setupvalue⚠

lua_status⚠

lua_toboolean⚠

lua_tocfunction⚠

lua_tointeger_⚠

lua_tolightuserdata⚠

lua_tolstring⚠

lua_tonumber⚠

lua_topointer⚠

lua_tostring⚠

lua_tothread⚠

lua_touserdata⚠

lua_type⚠

lua_typename⚠

lua_upvalueindex

lua_xmove⚠

lua_xpush⚠

lua_yield⚠

Type Aliases

lua_Alloc

Type for memory-allocation functions

lua_CFunction

Type for native C functions that can be passed to Lua.

lua_Hook

Type for functions to be called on debug events.

lua_Integer

lua_Number

A Lua number, usually equivalent to f64

lua_Reader

lua_Writer