mlua_sys::lua54

Module lua

Contains definitions from lua.h.

Structs

lua_Debug

lua_State

A raw Lua state associated with a thread.

Constants

LUAI_MAXSTACK

LUA_ERRERR

LUA_ERRMEM

LUA_ERRRUN

LUA_ERRSYNTAX

LUA_EXTRASPACE

LUA_GCCOLLECT

LUA_GCCOUNT

LUA_GCCOUNTB

LUA_GCGEN

LUA_GCINC

LUA_GCISRUNNING

LUA_GCRESTART

LUA_GCSETPAUSE

LUA_GCSETSTEPMUL

LUA_GCSTEP

LUA_GCSTOP

LUA_HOOKCALL

LUA_HOOKCOUNT

LUA_HOOKLINE

LUA_HOOKRET

LUA_HOOKTAILCALL

LUA_MASKCALL

LUA_MASKCOUNT

LUA_MASKLINE

LUA_MASKRET

LUA_MINSTACK

Minimum Lua stack available to a C function

LUA_MULTRET

LUA_NUMTYPES

LUA_OK

LUA_OPADD

LUA_OPBAND

LUA_OPBNOT

LUA_OPBOR

LUA_OPBXOR

LUA_OPDIV

LUA_OPEQ

LUA_OPIDIV

LUA_OPLE

LUA_OPLT

LUA_OPMOD

LUA_OPMUL

LUA_OPPOW

LUA_OPSHL

LUA_OPSHR

LUA_OPSUB

LUA_OPUNM

LUA_REGISTRYINDEX

LUA_RIDX_GLOBALS

LUA_RIDX_LAST

LUA_RIDX_MAINTHREAD

LUA_SIGNATURE

LUA_TBOOLEAN

LUA_TFUNCTION

LUA_TLIGHTUSERDATA

LUA_TNIL

LUA_TNONE

LUA_TNUMBER

LUA_TSTRING

LUA_TTABLE

LUA_TTHREAD

LUA_TUSERDATA

LUA_YIELD

Functions

lua_absindex⚠

lua_arith⚠

lua_atpanic⚠

lua_call⚠

lua_callk⚠

lua_checkstack⚠

lua_close⚠

lua_closeslot⚠

lua_compare⚠

lua_concat⚠

lua_copy⚠

lua_createtable⚠

lua_dump⚠

lua_error⚠

lua_gc⚠

lua_getallocf⚠

lua_getextraspace⚠

lua_getfield⚠

lua_getglobal⚠

lua_gethook⚠

lua_gethookcount⚠

lua_gethookmask⚠

lua_geti⚠

lua_getinfo⚠

lua_getiuservalue⚠

lua_getlocal⚠

lua_getmetatable⚠

lua_getstack⚠

lua_gettable⚠

lua_gettop⚠

lua_getupvalue⚠

lua_getuservalue⚠

lua_insert⚠

lua_isboolean⚠

lua_iscfunction⚠

lua_isfunction⚠

lua_isinteger⚠

lua_islightuserdata⚠

lua_isnil⚠

lua_isnone⚠

lua_isnoneornil⚠

lua_isnumber⚠

lua_isstring⚠

lua_istable⚠

lua_isthread⚠

lua_isuserdata⚠

lua_isyieldable⚠

lua_len⚠

lua_load⚠

lua_newstate⚠

lua_newtable⚠

lua_newthread⚠

lua_newuserdata⚠

lua_newuserdatauv⚠

lua_next⚠

lua_pcall⚠

lua_pcallk⚠

lua_pop⚠

lua_pushboolean⚠

lua_pushcclosure⚠

lua_pushcfunction⚠

lua_pushfstring⚠

lua_pushglobaltable⚠

lua_pushinteger⚠

lua_pushlightuserdata⚠

lua_pushliteral⚠

lua_pushlstring⚠

lua_pushnil⚠

lua_pushnumber⚠

lua_pushstring⚠

lua_pushthread⚠

lua_pushvalue⚠

lua_rawequal⚠

lua_rawget⚠

lua_rawgeti⚠

lua_rawgetp⚠

lua_rawlen⚠

lua_rawset⚠

lua_rawseti⚠

lua_rawsetp⚠

lua_register⚠

lua_remove⚠

lua_replace⚠

lua_resetthread⚠

lua_resume⚠

lua_rotate⚠

lua_setallocf⚠

lua_setfield⚠

lua_setglobal⚠

lua_sethook⚠

lua_seti⚠

lua_setiuservalue⚠

lua_setlocal⚠

lua_setmetatable⚠

lua_settable⚠

lua_settop⚠

lua_setupvalue⚠

lua_setuservalue⚠

lua_setwarnf⚠

lua_status⚠

lua_stringtonumber⚠

lua_toboolean⚠

lua_tocfunction⚠

lua_toclose⚠

lua_tointeger⚠

lua_tointegerx⚠

lua_tolightuserdata⚠

lua_tolstring⚠

lua_tonumber⚠

lua_tonumberx⚠

lua_topointer⚠

lua_tostring⚠

lua_tothread⚠

lua_touserdata⚠

lua_type⚠

lua_typename⚠

lua_upvalueid⚠

lua_upvalueindex

lua_upvaluejoin⚠

lua_version⚠

lua_warning⚠

lua_xmove⚠

lua_xpush⚠

lua_yield⚠

lua_yieldk⚠

Type Aliases

lua_Alloc

Type for memory-allocation functions

lua_CFunction

Type for native C functions that can be passed to Lua

lua_Hook

Type for functions to be called on debug events.

lua_Integer

A Lua integer, usually equivalent to i64

lua_KContext

Type for continuation-function contexts

lua_KFunction

Type for continuation functions

lua_Number

A Lua number, usually equivalent to f64

lua_Reader

lua_Unsigned

A Lua unsigned integer, usually equivalent to u64

lua_WarnFunction

Type for warning functions

lua_Writer