mlua_sys::luau

Module lua

Contains definitions from lua.h.

Structs

lua_Callbacks

lua_Debug

lua_State

A raw Lua state associated with a thread.

Constants

LUA_ENVIRONINDEX

LUA_ERRERR

LUA_ERRMEM

LUA_ERRRUN

LUA_ERRSYNTAX

LUA_GCCOLLECT

LUA_GCCOUNT

LUA_GCCOUNTB

LUA_GCISRUNNING

LUA_GCRESTART

LUA_GCSETGOAL

LUA_GCSETSTEPMUL

LUA_GCSETSTEPSIZE

LUA_GCSTEP

LUA_GCSTOP

LUA_GLOBALSINDEX

LUA_MINSTACK

Guaranteed number of Lua stack slots available to a C function.

LUA_MULTRET

LUA_NOREF

LUA_OK

LUA_REFNIL

LUA_REGISTRYINDEX

LUA_TBOOLEAN

LUA_TBUFFER

LUA_TFUNCTION

LUA_TLIGHTUSERDATA

LUA_TNIL

LUA_TNONE

LUA_TNUMBER

LUA_TSTRING

LUA_TTABLE

LUA_TTHREAD

LUA_TUSERDATA

LUA_TVECTOR

LUA_YIELD

Functions

lua_absindex⚠

lua_break⚠

lua_breakpoint⚠

lua_call⚠

lua_callbacks⚠

lua_checkstack⚠

lua_cleartable⚠

lua_clock⚠

lua_clonefunction⚠

lua_close⚠

lua_concat⚠

lua_createtable⚠

lua_debugtrace⚠

lua_equal⚠

lua_error⚠

lua_gc⚠

lua_getallocf⚠

lua_getargument⚠

lua_getcoverage⚠

lua_getfenv⚠

lua_getfield⚠

lua_getglobal⚠

lua_getinfo⚠

lua_getlightuserdataname⚠

lua_getlocal⚠

lua_getmetatable⚠

lua_getmetatablepointer⚠

lua_getreadonly⚠

lua_gettable⚠

lua_getthreaddata⚠

lua_gettop⚠

lua_getupvalue⚠

lua_getuserdatadtor⚠

lua_getuserdatametatable⚠

lua_insert⚠

lua_isLfunction⚠

lua_isboolean⚠

lua_isbuffer⚠

lua_iscfunction⚠

lua_isfunction⚠

lua_islightuserdata⚠

lua_isnil⚠

lua_isnone⚠

lua_isnoneornil⚠

lua_isnumber⚠

lua_isstring⚠

lua_istable⚠

lua_isthread⚠

lua_isthreadreset⚠

lua_isuserdata⚠

lua_isvector⚠

lua_isyieldable⚠

lua_lessthan⚠

lua_lightuserdatatag⚠

lua_mainthread⚠

lua_namecallatom⚠

lua_newbuffer⚠

lua_newstate⚠

lua_newtable⚠

lua_newthread⚠

lua_newuserdata⚠

lua_newuserdata_t⚠

lua_newuserdatadtor⚠

lua_newuserdatatagged⚠

lua_newuserdatataggedwithmetatable⚠

lua_next⚠

lua_objlen⚠

lua_pcall⚠

lua_pop⚠

lua_pushboolean⚠

lua_pushcclosure⚠

lua_pushcclosured⚠

lua_pushcclosurek⚠

lua_pushcfunction⚠

lua_pushcfunctiond⚠

lua_pushfstring⚠

lua_pushinteger⚠

lua_pushlightuserdata⚠

lua_pushlightuserdatatagged⚠

lua_pushliteral⚠

lua_pushlstring_⚠

lua_pushnil⚠

lua_pushnumber⚠

lua_pushstring_⚠

lua_pushthread⚠

lua_pushunsigned⚠

lua_pushvalue⚠

lua_pushvector⚠

lua_rawcheckstack⚠

lua_rawequal⚠

lua_rawget⚠

lua_rawgetfield⚠

lua_rawgeti_⚠

lua_rawiter⚠

lua_rawset⚠

lua_rawseti_⚠

lua_ref⚠

lua_remove⚠

lua_replace⚠

lua_resetthread⚠

lua_resume_⚠

lua_resumeerror⚠

lua_setfenv⚠

lua_setfield⚠

lua_setglobal⚠

lua_setlightuserdataname⚠

lua_setlocal⚠

lua_setmemcat⚠

lua_setmetatable⚠

lua_setreadonly⚠

lua_setsafeenv⚠

lua_settable⚠

lua_setthreaddata⚠

lua_settop⚠

lua_setupvalue⚠

lua_setuserdatadtor⚠

lua_setuserdatametatable⚠

lua_setuserdatatag⚠

lua_singlestep⚠

lua_stackdepth⚠

lua_status⚠

lua_toboolean⚠

lua_tobuffer⚠

lua_tocfunction⚠

lua_tointeger_⚠

lua_tointegerx_⚠

lua_tolightuserdata⚠

lua_tolightuserdatatagged⚠

lua_tolstring⚠

lua_tonumber⚠

lua_tonumberx⚠

lua_topointer⚠

lua_tostring⚠

lua_tostringatom⚠

lua_totalbytes⚠

lua_tothread⚠

lua_tounsigned⚠

lua_tounsignedx⚠

lua_touserdata⚠

lua_touserdatatagged⚠

lua_tovector⚠

lua_type⚠

lua_typename⚠

lua_unref⚠

lua_upvalueindex

lua_userdatatag⚠

lua_xmove⚠

lua_xpush⚠

lua_yield⚠

luau_load⚠

luau_setfflag⚠

luau_version

Returns Luau release version (eg. 0.xxx).

Type Aliases

lua_Alloc

Type for memory-allocation functions.

lua_CFunction

Type for native C functions that can be passed to Lua.

lua_Continuation

lua_Coverage

lua_Destructor

lua_Hook

Type for functions to be called on debug events.

lua_Integer

A Lua integer, equivalent to i32.

lua_Number

A Lua number, usually equivalent to f64.

lua_Udestructor

Type for userdata destructor functions.

lua_Unsigned

A Lua unsigned integer, equivalent to u32.